Solve the problem
오류 해결했다. 방법은 script.js에서 require 받지말고 server에서 데이터를
전처리해서 script.js 파일로 전송!
