+++ public/index.html
@@ -21,7 +21,7 @@
     />
 
     <link rel="stylesheet" href="style.css" />
-    <script defer src="./script.js" type="module"></script>
+    <script defer src="./script.js"></script>
   </head>
 
   <body>
