+++ server.js
@@ -20,6 +20,11 @@ app.use(express.static('public'));
 app.use(express.json());
 
 
+const {initialize} = require("koalanlp/Util");
+const {Tagger} = require("koalanlp/proc");
+const {EUNJEON} = require("koalanlp/API");
+
+
 app.get("/",(req, res) => {
   // root url, 즉 메인 페이지로 접속했을 때 papago의 메인 페이지가 나와야함.
   // public / ~
@@ -47,13 +52,103 @@ app.post("/news", (req, res) => {
 
   request.get(options, (error, response, body) => { 
     if(!error && response.statusCode == 200){
+      // const parseBody = JSON.parse(body); // parse() : string -> object로 변환
+      // console.log(typeof parseBody, parseBody);
 
-      res.json(body);
+      const parseJsonToObject = JSON.parse(body);
+      console.log(typeof parseJsonToObject, parseJsonToObject);
+
+      const display = parseJsonToObject['display']; 
+      console.log(display);
+
+      var description;
+      for(var i = 0; i < display ; i++){
+        description = description + parseJsonToObject['items'][i]['description'];
+      }
+
+      console.log(description);
+
+      
+
+      var sortable = {};
+
+      initialize({packages: {EUNJEON: 'LATEST'}}).then(() => {
+        let tagger = new Tagger(EUNJEON);
+        const text = description; 
+        // let tagged = await tagger(text);
+        
+        var arrNumber = new Array();
+        var index = 0;
+        let tagged = tagger.tagSync(text);
+          tagged.forEach((sent, i) => {
+              console.log(`===== Sentence #${i} =====`);
+              console.log(sent.surfaceString());
+      
+              console.log("# Analysis Result");
+              // console.log(sent.singleLineString());
+              sent.forEach((word) => {
+                    // console.log(`Word [${word.id}] ${word.surface} = `);
+      
+                  // word.forEach((morph) => {
+                  //     console.log(`${morph.surface}/${morph.tag} `);
+                  // });
+                  
+                  word.forEach((morph) =>{
+                    if(morph.tag['tagname'] == 'NNP' || morph.tag['tagname'] == 'NNG'){
+                      // console.log(morph.surface);
+                      arrNumber[index] = morph.surface;
+                      // console.log(index);
+                      // console.log(morph.surface);
+                      index++;
+                    }
+                  })
+      
+              });
+          });
+      
+          const result = {};
+          arrNumber.forEach((x) => { 
+            result[x] = (result[x] || 0)+1; 
+          });
+          
+      
+          sortable = Object.entries(result)
+          .sort(([, a], [, b]) => b - a)
+          .reduce((r, [k, v]) => ({ ...r, [k]: v }), {});
+      
+          console.log(typeof sortable, sortable);
+          
+          // console.log(JSON.stringify(sortable));
+          
+          // console.log(Object.keys(sortable));
+          
+          // console.log(Object.keys.sortable[0])
+      
+          // var index_r = 0;
+          // for (i of Object.keys(sortable)) {
+          //   arrResult[index_r] = i; 
+          //   index_r++;
+          // }
+          res.json(sortable);
+          console.log('finished!');
+        
+      });
+
+
+      // console.log(Object.keys.sortable[0])
+
+      // res.redirect('keyword');
+      // console.log(typeof sortable, sortable, "dsdsdsdsssssssssssssssssssd");
+      // res.json(sortable);
     }else{
       console.log(`error = ${response.statusCode}`);
     }
   });
 });
+
+
+
+
